Fix for reconnect stale data issues
As Bybit still do not send through snapshot data for private topics, if the WS encounters an error (eg. ping/pong timed out) and restart_on_error (self.handle_error) is set to True, the WS will attempt to reconnect.

If successful reconnect, the self.data dictionary entries for private topics can/will become stale in the time it takes to reconnect, as there is no snapshot on connect to overwrite with current data.

Public topic subscriptions will be overwritten by the new snapshot data, so it is being cleared/updated in any case.

This fix adds a common reset function for the state booleans and to clear the (potentially stale) dictionary.